+++ packages/gatsby-admin/gatsby-config.js
@@ -1,4 +1,16 @@
 module.exports = {
-  plugins: [`gatsby-plugin-react-helmet`],
+  plugins: [{
+    resolve: "gatsby-plugin-react-helmet"
+  }, {
+    resolve: "gatsby-plugin-webfonts",
+    options: {
+      fonts: {
+        google: [{
+          family: `Inter`,
+          variants: [`400`, `600`, `800`]
+        }]
+      }
+    }
+  }],
   pathPrefix: `/___admin`
 };