+++ packages/gatsby-recipes/src/providers/npm/package-json.test.js
@@ -34,6 +34,36 @@ describe(`packageJson resource`, () => {
     })
   })
 
+  test(`handles multiple parallel create calls`, async () => {
+    const resultPromise = pkgJson.create(
+      {
+        root,
+      },
+      {
+        name: `husky`,
+        value: JSON.parse(initialValue),
+      }
+    )
+    const result2Promise = pkgJson.create(
+      {
+        root,
+      },
+      {
+        name: `husky2`,
+        value: JSON.parse(initialValue),
+      }
+    )
+
+    const result = await resultPromise
+    const result2 = await result2Promise
+
+    expect(result).toMatchSnapshot()
+    expect(result2).toMatchSnapshot()
+
+    await pkgJson.destroy({ root }, result)
+    await pkgJson.destroy({ root }, result2)
+  })
+
   test(`handles object values`, async () => {
     const result = await pkgJson.create(
       {