+++ packages/gatsby/src/services/calculate-dirty-queries.ts
@@ -1,18 +1,24 @@
-import { calcInitialDirtyQueryIds, groupQueryIds } from "../query"
+import {
+  calcInitialDirtyQueryIds,
+  calcDirtyQueryIds,
+  groupQueryIds,
+} from "../query"
 import { IGroupedQueryIds } from "./"
 import { IQueryRunningContext } from "../state-machines/query-running/types"
 import { assertStore } from "../utils/assert-store"
 
 export async function calculateDirtyQueries({
   store,
+  firstRun,
 }: Partial<IQueryRunningContext>): Promise<{
   queryIds: IGroupedQueryIds
 }> {
   assertStore(store)
 
   const state = store.getState()
-  // TODO: Check filesDirty from context
 
-  const queryIds = calcInitialDirtyQueryIds(state)
+  const queryIds = firstRun
+    ? calcInitialDirtyQueryIds(state)
+    : calcDirtyQueryIds(state)
   return { queryIds: groupQueryIds(queryIds) }
 }