+++ packages/gatsby-plugin-styletron/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styletron@5.0.8...gatsby-plugin-styletron@5.0.9) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-plugin-styletron
+
 ## [5.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-styletron@5.0.7...gatsby-plugin-styletron@5.0.8) (2020-07-02)
 
 **Note:** Version bump only for package gatsby-plugin-styletron