+++ packages/gatsby-source-drupal/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.26...gatsby-source-drupal@3.5.27) (2020-08-05)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.5.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.25...gatsby-source-drupal@3.5.26) (2020-07-24)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.5.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.24...gatsby-source-drupal@3.5.25) (2020-07-22)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.5.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.23...gatsby-source-drupal@3.5.24) (2020-07-21)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.5.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.22...gatsby-source-drupal@3.5.23) (2020-07-17)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.5.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.21...gatsby-source-drupal@3.5.22) (2020-07-15)
+
+### Features
+
+- **gatsby-source-drupal:** Add Drupal Fastbuilds ([#25428](https://github.com/gatsbyjs/gatsby/issues/25428)) ([64685c3](https://github.com/gatsbyjs/gatsby/commit/64685c3))
+
+## [3.5.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.20...gatsby-source-drupal@3.5.21) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
 ## [3.5.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.5.19...gatsby-source-drupal@3.5.20) (2020-07-02)
 
 **Note:** Version bump only for package gatsby-source-drupal