+++ packages/gatsby-transformer-csv/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-csv@2.3.9...gatsby-transformer-csv@2.3.10) (2020-07-09)
+
+**Note:** Version bump only for package gatsby-transformer-csv
+
 ## [2.3.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-csv@2.3.8...gatsby-transformer-csv@2.3.9) (2020-07-02)
 
 **Note:** Version bump only for package gatsby-transformer-csv